Make telemetry truly optional
